Code/python-code/requirements.txt
pyopengl 
glfw 
pyglm 
pillow 
numpy-stl
pygltf
